src/with_naver_api/requirements.txt
langchain-mcp-adapters==0.0.5
langchain-openai==0.3.11
langgraph-api==0.0.38
langgraph-cli==0.1.81
pip==25.0.1
xmltodict==0.14.2
